+++ internal/splith1headers/config.go
@@ -5,15 +5,12 @@ var constants = struct {
 	outputFolder      string
 	notes             string
 	href              string
-	// Set `maxElemsPerFile` to -1 to have no limit
-	maxElemsPerFile   int
 	primaryHeadersMap map[string]struct{}
 }{
-	baseFileName:    "Body",
-	outputFolder:    "output",
-	notes:           "Notes",
-	href:            "href",
-	maxElemsPerFile: -1,
+	baseFileName: "Body",
+	outputFolder: "output",
+	notes:        "Notes",
+	href:         "href",
 	primaryHeadersMap: map[string]struct{}{
 		"h1": {},
 		"h2": {},